requirements.txt
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
python-dotenv>=1.0.0
tqdm>=4.65.0
streamlit>=1.28.0

